pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ecsify"
version = "0.1.0"
description = "A CLI tool to simplify deployment of containers to AWS ECS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "ECSify Contributors", email = "contributors@ecsify.dev" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "boto3>=1.28.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "types-PyYAML",
]

[project.scripts]
ecsify = "ecsify.cli:main"

[project.urls]
Homepage = "https://github.com/sanchezpaco/ecsify"
Repository = "https://github.com/sanchezpaco/ecsify.git"
Issues = "https://github.com/sanchezpaco/ecsify/issues"

[tool.hatch.build.targets.wheel]
packages = ["ecsify"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["ecsify"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
